+++ gettsim/tests/test_vorsorge.py
@@ -59,5 +59,5 @@ def test_vorsorge(
 
     # TODO: Here our test values are off by about 5 euro. We should revisit. See #217.
     assert_series_equal(
-        result[target], year_data[target], check_less_precise=1, check_dtype=False
+        result[target], year_data[target], atol=1e-1, rtol=1, check_dtype=False
     )