+++ CHANGES.rst
@@ -5,7 +5,12 @@ This is a record of all past ``gettsim`` releases and what went into them in rev
 chronological order. We follow `semantic versioning <https://semver.org/>`_ and all
 releases are available on `Anaconda.org <https://anaconda.org/gettsim/gettsim>`_.
 
-0.4.x -
+0.4.2 -
+------------------
+* :gh:`285` Set up bokeh dashboard to visualize tax-benefit parameters.
+   (:ghuser:`Si-Pf`, :ghuser:`Eric-Sommer`).
+
+0.4.1 - 2021-04-11
 ------------------
 
 * :gh:`248` Implement 2021 increase in lump sum tax deductions for disabled.
@@ -23,7 +28,7 @@ releases are available on `Anaconda.org <https://anaconda.org/gettsim/gettsim>`_
 * :gh:`235` Module to create synthetic data
     (:ghuser:`Eric-Sommer`, :ghuser:`ChristianZimpelmann`).
 * :gh:`256` Implement 2021 housing benefit reform (:ghuser:`Eric-Sommer`).
-
+* :gh:`283` Implement Kinderbonus in 2020 and 2021 (:ghuser:`ChristianZimpelmann`).
 
 
 0.4.0 - 2020-11-11